Added Dockerfile in repo root.
Need to make it available outside the docker folder as the build context was specified in the docker compose as the repo root directory and corresponding paths in the Dockerfile were relative to the root directory.

Additionally, added Environment variables defined in docker compose file.

The image wasn't running directly after building because these things were missing:
- needed to add db container as well
- ports needed to be exposed for both db and dashboard container
- common network needed to be added under which both the containers had to be linked
- all environment variables had to be added, especially important ones like DB_HOST