Replaced config.py with config-fake.py in external Dockerfile.
Encountered error while building and pushing docker image as a part of the CI.
Due to config.py mentioned in COPY command in Dockerfile not being there in the repo online.
Added config-fake.py in its place for now, so image is pushed to Dockerhub.
But eventually, will need to handle as per decision to consolidate differences in external and internal repos.