Replacing CognitoConfig class with ENV variables
- Option 1: Can we read in these values directly from environment variables instead of reading from a class?

- Option 2: Create a class that makes all environment variables available as python variables

For now, I’m going ahead with reading in these variables directly.
So, will need to replace all uses of CognitoConfig class with variable name directly in files that use it.

Shankari mentioned in a commit whether we even need config.py?
https://github.com/e-mission/op-admin-dashboard/commit/927817a95a9f7f9c4307f98cbfe2ef02584c047d

Also, removed references to both config-fake.py and config.py in the Dockerfiles.