Changing base image to build from redesign server image
Created a new branch image-push-merge for e-mission-server in my forked repo.
Also modified image push workflow to build and push docker image to docker hub on push to image-push-merge branch.

Doing this since admin-dash was failing when was building from internal repo since this was still referring to old server code.
Changed Dockerfile and docker/Dockerfile.dev in public-dash and admin-dash to build from this new image
mukuflash03/e-mission-server:image-push-merge_2024-04-12--01-01

Redesigned server image is built from the image-push-merge branch on my personal forked repository.
This branch has a workflow run set up to build the docker image and push it to Dockerhub whenever a push or merge happens to image-push-merge branch.
Currently, I've been pushing to image-push and then creating a PR to merge into image-push-merge.

Doing this, so admin-dash and public-dash can build from the latest redesigned server code.
This is not the latest server code but the latest changes from my redesign PR.