Add performance stats to the admin dashboard (#141)
Use the new `store_dashboard_time` function from https://github.com/e-mission/e-mission-server to instrument sections of the admin dashboard codebase, both in summary and in detail. We will use this to document the effect of our scalability improvements.

TODO/Future fix: Revisit the method in which we instrument to be less repetitive by using a decorator or enhancing the timer. https://github.com/e-mission/op-admin-dashboard/pull/141#issuecomment-2424227569